Reorganizing widgets and pages structure as good practice
