+++ lib/widgets/search_bar.dart
