+++ lib/widgets/search_result.dart
